+++ Tidify/Targets/TidifyPresentation/Sources/Common/Base/BaseViewController.swift
@@ -14,12 +14,6 @@ class BaseViewController: UIViewController {
   // MARK: Properties
   var cancellable: Set<AnyCancellable> = []
 
-  lazy var indicatorView: UIActivityIndicatorView = {
-    let indicatorView: UIActivityIndicatorView = .init()
-    indicatorView.color = .t_blue()
-    return indicatorView
-  }()
-
   // MARK: LifeCycle
   override func viewDidLoad() {
     super.viewDidLoad()
@@ -30,13 +24,5 @@ class BaseViewController: UIViewController {
   // MARK: Methods
   func setupViews() {
     view.backgroundColor = .t_background()
-  }
-
-  func setIndicatorView(isLoading: Bool) {
-    if isLoading {
-      indicatorView.startAnimating()
-    } else {
-      indicatorView.isHidden = true
-    }
   }
 }
